Implemented basic group features
Groups can be created. Creator will be added as user. Invitations are not implemented. Group profiles will show all messages sent to and by users in the group, as well as show a list of the users in the group. The generic /groups page will show a list of groups the currently logged in user is a part of. This also seems like the logical place to handle invitations and the like