We can now attach graphs to WallPost.
WallPosts (when posted on the user's /wall page) can have a graph attached to them. They graph can be picked from a list (probably need to work on front-end there, it's not pretty), and then gets added to the message. Posting the message will show a small version of the image, until you click on it, when it will become bigger. Probably possible to allow constant switching between sizes, but I'm tired.

Making the graph list requires the user having saved a graph, which you can do on /dashboard. Underneath the graph element, there's a textfield in which you input a unique name for the graph, the name that will be shown in the list, and then when you hit save, it will be posted to the server and stored.

I think that's it. Some extra database stuff, Model, API, but all of it seemed to work as far as I could tell.

No thorough testing, so putting this up on seperate branch.